Lag utility-fil og legg finnStrengesteDiskresjonsKode() til denne
