+++ src/main/kotlin/no/nav/familie/ba/sak/behandling/ArbeidsfordelingService.kt
@@ -24,7 +24,7 @@ class ArbeidsfordelingService(val behandlingRepository: BehandlingRepository,
             barn -> integrasjonTjeneste.hentPersoninfoFor(barn.personIdent.ident)
         }.plus(søker)
 
-        val strengesteDiskresjonskode = DiskresjonskodeService.finnStrengesteDiskresjonskode(personinfoliste)
+        val strengesteDiskresjonskode = finnStrengesteDiskresjonskode(personinfoliste)
 
         return integrasjonTjeneste.hentBehandlendeEnhet(søker.geografiskTilknytning, strengesteDiskresjonskode)
     }
